Changed 'is' function to handle Except; Moved 'is' function to ops
