+++ runtime/su.ts
@@ -14,11 +14,10 @@ import { SuRecord } from "./surecord";
 import { SuDate } from "./sudate";
 import { type } from "./type";
 import { display } from "./display";
-import { is } from "./is";
 import { cmp } from "./cmp";
 import { global } from "./global";
 import { Strings } from "./builtin/strings";
-import { isString, coerceStr, toStr } from "./ops";
+import { isString, coerceStr, toStr, is } from "./ops";
 const sm: any = Strings.prototype;
 import { Numbers } from "./builtin/numbers";
 const nm: any = Numbers.prototype;
